+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Added
 - button copy link to editor, in the table with files
+- ability to insert a link to the editor on the page
 - access for an anonymous user, open the document for viewing in the editor if the document is attached to a public post
 
 ## Changed