+++ CHANGELOG.md
@@ -7,10 +7,13 @@
 - button copy link to editor, in the table with files
 - ability to insert a link to the editor on the page
 - access for an anonymous user, open the document for viewing in the editor if the document is attached to a public post
+- filling pdf
 
 ## Changed
 - editor url
 - displaying the onlyoffice-wordpress-block component in the Gutenberg editor
+- remove filling for oform
+- supported formats updated
 
 ## 1.1.0
 ## Added